ui(refactor): split 'superdesk-desks' module.
