+++ scripts/superdesk-desks/constants.js
@@ -1,0 +1,4 @@
+export const limits = {
+    stage: 15,
+    desk: 40
+};
