+++ scripts/superdesk-desks/controllers/DeskSettingsController.js
@@ -1,0 +1,7 @@
+DeskSettingsController.$inject = ['$scope', 'desks'];
+export function DeskSettingsController($scope, desks) {
+    desks.initialize()
+    .then(function() {
+        $scope.desks = desks.desks;
+    });
+}
