+++ scripts/superdesk-desks/controllers/index.js
@@ -1,0 +1,3 @@
+export { DeskListController } from './DeskListController';
+export { DeskSettingsController } from './DeskSettingsController';
+export { DeskConfigController } from './DeskConfigController';
