+++ scripts/superdesk-desks/directives/SluglinesItemListDirective.js
@@ -1,0 +1,18 @@
+SluglinesItemListDirective.$inject = ['api'];
+export function SluglinesItemListDirective(api) {
+    return {
+        templateUrl: 'scripts/superdesk-desks/views/slugline-items.html',
+        scope: {
+            desk: '='
+        },
+        link: function(scope, elem) {
+            scope.items = [];
+            scope.loading = true;
+            api.get('desks/' + scope.desk + '/sluglines').then(function(items) {
+                scope.items = items._items;
+            })['finally'](function() {
+                scope.loading = false;
+            });
+        }
+    };
+}
