+++ scripts/superdesk-desks/directives/StageHeaderDirective.js
@@ -1,0 +1,3 @@
+export function StageHeaderDirective() {
+    return {templateUrl: 'scripts/superdesk-desks/views/stage-header.html'};
+}
