+++ scripts/superdesk-desks/services/index.js
@@ -1,0 +1,1 @@
+export { DesksFactory } from './DesksFactory';
